+++ src/flows/Onboarding/api.ts
@@ -1,0 +1,340 @@
+import {
+  Employment,
+  EmploymentFullParams,
+  getIndexBenefitOffer,
+  getShowCompany,
+  getShowEmployment,
+  getShowFormCountry,
+  getShowSchema,
+  patchUpdateEmployment2,
+  postInviteEmploymentInvitation,
+  PostInviteEmploymentInvitationData,
+  putUpdateBenefitOffer,
+  UnifiedEmploymentUpsertBenefitOffersRequest,
+} from '@/src/client';
+import { useClient } from '@/src/context';
+import { OnboardingFlowParams } from '@/src/flows/Onboarding/types';
+import { JSONSchemaFormType } from '@/src/flows/types';
+import { Client } from '@hey-api/client-fetch';
+import { createHeadlessForm, modify } from '@remoteoss/json-schema-form';
+import { useMutation, useQuery } from '@tanstack/react-query';
+import { FieldValues } from 'react-hook-form';
+
+const jsonSchemaToEmployment: Partial<
+  Record<JSONSchemaFormType, keyof Employment>
+> = {
+  employment_basic_information: 'basic_information',
+  contract_details: 'contract_details',
+};
+
+export const useEmployment = (employmentId: string | undefined) => {
+  const { client } = useClient();
+
+  return useQuery({
+    queryKey: ['employment', employmentId],
+    retry: false,
+    enabled: !!employmentId,
+    queryFn: async () => {
+      const response = await getShowEmployment({
+        client: client as Client,
+        headers: {
+          Authorization: ``,
+        },
+        path: {
+          employment_id: employmentId as string,
+        },
+      });
+
+      // If response status is 404 or other error, throw an error to trigger isError
+      if (response.error || !response.data) {
+        throw new Error('Failed to fetch employment data');
+      }
+
+      return response;
+    },
+  });
+};
+
+export const useCompany = (companyId: string) => {
+  const { client } = useClient();
+  return useQuery({
+    queryKey: ['company', companyId],
+    retry: false,
+    enabled: !!companyId,
+    queryFn: async () => {
+      const response = await getShowCompany({
+        client: client as Client,
+        headers: {
+          Authorization: ``,
+        },
+        path: {
+          company_id: companyId,
+        },
+      });
+
+      if (response.error || !response.data) {
+        throw new Error('Failed to fetch employment data');
+      }
+
+      return response;
+    },
+    select: ({ data }) => {
+      return data.data.company;
+    },
+  });
+};
+
+export const useBenefitOffers = (employmentId: string | undefined) => {
+  const { client } = useClient();
+  return useQuery({
+    queryKey: ['benefit-offers', employmentId],
+    retry: false,
+    enabled: !!employmentId,
+    queryFn: async () => {
+      return getIndexBenefitOffer({
+        client: client as Client,
+        headers: {
+          Authorization: ``,
+        },
+        path: {
+          employment_id: employmentId as string,
+        },
+      }).then((response) => {
+        // If response status is 404 or other error, throw an error to trigger isError
+        if (response.error || !response.data) {
+          throw new Error('Failed to fetch benefit offers data');
+        }
+
+        return response;
+      });
+    },
+    select: ({ data }) =>
+      data?.data?.reduce(
+        (acc, item) => {
+          return {
+            ...acc,
+            [item.benefit_group.slug]: {
+              value: item.benefit_tier?.slug ?? '',
+            },
+          };
+        },
+        {} as Record<string, { value: string }>,
+      ),
+  });
+};
+/**
+ * Use this hook to invite an employee to the onboarding flow
+ * @returns
+ */
+export const useEmploymentInvite = () => {
+  const { client } = useClient();
+
+  return useMutation({
+    mutationFn: (payload: PostInviteEmploymentInvitationData['path']) => {
+      return postInviteEmploymentInvitation({
+        client: client as Client,
+        headers: {
+          Authorization: ``,
+        },
+        path: payload,
+      });
+    },
+  });
+};
+
+// TODO: This hooks uses the incorrect endpoint as it hasn't been created yet.``
+export const useCreateReserve = () => {
+  //const { client } = useClient();
+  return useMutation({
+    mutationFn: (payload: { employment_id: string }) => {
+      return Promise.resolve(payload);
+    },
+  });
+};
+
+/**
+ * Use this hook to get the JSON schema form for the onboarding flow
+ * @param param0
+ * @returns
+ */
+export const useJSONSchemaForm = ({
+  countryCode,
+  form,
+  fieldValues,
+  options,
+  employment,
+}: {
+  countryCode: string;
+  form: JSONSchemaFormType;
+  fieldValues: FieldValues;
+  options?: OnboardingFlowParams['options'];
+  employment?: Employment;
+}) => {
+  const { client } = useClient();
+  const jsonSchemaQueryParam = options?.jsonSchemaVersion?.form_schema?.[form]
+    ? {
+        json_schema_version: options.jsonSchemaVersion.form_schema[form],
+      }
+    : {};
+  return useQuery({
+    queryKey: ['onboarding-json-schema-form', countryCode, form],
+    retry: false,
+    queryFn: async () => {
+      const response = await getShowFormCountry({
+        client: client as Client,
+        headers: {
+          Authorization: ``,
+        },
+        path: {
+          country_code: countryCode,
+          form: form,
+        },
+        query: {
+          skip_benefits: true,
+          ...jsonSchemaQueryParam,
+        },
+      });
+
+      // If response status is 404 or other error, throw an error to trigger isError
+      if (response.error || !response.data) {
+        throw new Error('Failed to fetch onboarding schema');
+      }
+
+      return response;
+    },
+    select: ({ data }) => {
+      let jsfSchema = data?.data || {};
+      if (options && options.jsfModify) {
+        const { schema } = modify(jsfSchema, options.jsfModify);
+        jsfSchema = schema;
+      }
+      const hasFieldValues = Object.keys(fieldValues).length > 0;
+      const employmentField = jsonSchemaToEmployment[form] as keyof Employment;
+      const employmentFieldData = (employment?.[employmentField] ||
+        {}) as Record<string, unknown>;
+      return createHeadlessForm(jsfSchema, {
+        initialValues: hasFieldValues ? fieldValues : employmentFieldData,
+      });
+    },
+  });
+};
+
+export const useBenefitOffersSchema = (
+  employmentId: string,
+  fieldValues: FieldValues,
+  options: OnboardingFlowParams['options'],
+) => {
+  const jsonSchemaQueryParam = options?.jsonSchemaVersion
+    ?.benefit_offers_form_schema
+    ? {
+        json_schema_version:
+          options.jsonSchemaVersion.benefit_offers_form_schema,
+      }
+    : {};
+  const { client } = useClient();
+  return useQuery({
+    queryKey: ['benefit-offers-schema', employmentId],
+    retry: false,
+    enabled: !!employmentId,
+    queryFn: async () => {
+      const response = await getShowSchema({
+        client: client as Client,
+        headers: {
+          Authorization: ``,
+        },
+        path: {
+          employment_id: employmentId,
+        },
+        query: jsonSchemaQueryParam,
+      });
+
+      // If response status is 404 or other error, throw an error to trigger isError
+      if (response.error || !response.data) {
+        throw new Error('Failed to fetch benefit offers schema');
+      }
+
+      return response;
+    },
+    select: ({ data }) => {
+      let jsfSchema = data?.data?.schema || {};
+
+      if (options && options.jsfModify) {
+        const { schema } = modify(jsfSchema, options.jsfModify);
+        jsfSchema = schema;
+      }
+      const hasFieldValues = Object.keys(fieldValues).length > 0;
+      const result = createHeadlessForm(jsfSchema, {
+        // we need to clone the fieldValues to prevent side effects
+        // if we don't do this, the benefits get included in the other steps
+        initialValues: hasFieldValues ? { ...fieldValues } : {},
+      });
+      return result;
+    },
+  });
+};
+
+/**
+ * Use this hook to create an employment
+ * @returns
+ */
+export const useCreateEmployment = () => {
+  const { client } = useClient();
+  return useMutation({
+    mutationFn: (payload: EmploymentCreateParams) => {
+      return postCreateEmployment2({
+        client: client as Client,
+        headers: {
+          Authorization: ``,
+        },
+        body: payload,
+      });
+    },
+  });
+};
+
+export const useUpdateEmployment = () => {
+  const { client } = useClient();
+  return useMutation({
+    mutationFn: ({
+      employmentId,
+      ...payload
+    }: EmploymentFullParams & { employmentId: string }) => {
+      return patchUpdateEmployment2({
+        client: client as Client,
+        headers: {
+          Authorization: ``,
+        },
+        body: payload,
+        path: {
+          employment_id: employmentId,
+        },
+        query: {
+          skip_benefits: true,
+        },
+      });
+    },
+  });
+};
+
+export const useUpdateBenefitsOffers = () => {
+  const { client } = useClient();
+  return useMutation({
+    mutationFn: ({
+      employmentId,
+      ...payload
+    }: UnifiedEmploymentUpsertBenefitOffersRequest & {
+      employmentId: string;
+    }) => {
+      return putUpdateBenefitOffer({
+        client: client as Client,
+        headers: {
+          Authorization: ``,
+        },
+        body: payload,
+        path: {
+          employment_id: employmentId,
+        },
+      });
+    },
+  });
+};
